abdm m2 and m3 flow

+++ backend/src/main/java/com/hadproject/dhanvantari/visit/Visit.java
@@ -30,7 +30,7 @@ public class Visit {
             strategy = GenerationType.SEQUENCE,
             generator = "visit_sequence"
     )
-    private int visitId;
+    private Long visitId;
 
     @Column
     private String prescription;
